Apply API review feedback
