+++ src/Components/Endpoints/src/PublicAPI.Unshipped.txt
@@ -41,6 +41,18 @@ Microsoft.AspNetCore.Components.Endpoints.RazorComponentResult<TComponent>.Razor
 Microsoft.AspNetCore.Components.Endpoints.RazorComponentResult<TComponent>.RazorComponentResult(System.Collections.Generic.IReadOnlyDictionary<string!, object?>! parameters) -> void
 Microsoft.AspNetCore.Components.Endpoints.RazorComponentResultExecutor
 Microsoft.AspNetCore.Components.Endpoints.RazorComponentResultExecutor.RazorComponentResultExecutor() -> void
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.FormMappingUseCurrentCulture.get -> bool
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.FormMappingUseCurrentCulture.set -> void
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.MaxFormMappingCollectionSize.get -> int
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.MaxFormMappingCollectionSize.set -> void
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.MaxFormMappingErrorCount.get -> int
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.MaxFormMappingErrorCount.set -> void
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.MaxFormMappingKeySize.get -> int
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.MaxFormMappingKeySize.set -> void
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.MaxFormMappingRecursionDepth.get -> int
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.MaxFormMappingRecursionDepth.set -> void
+Microsoft.AspNetCore.Components.Endpoints.RazorComponentsOptions.RazorComponentsOptions() -> void
 Microsoft.AspNetCore.Components.Endpoints.RenderModeEndpointProvider
 Microsoft.AspNetCore.Components.Endpoints.RenderModeEndpointProvider.RenderModeEndpointProvider() -> void
 Microsoft.AspNetCore.Components.Endpoints.RootComponentMetadata
@@ -54,24 +66,21 @@ Microsoft.AspNetCore.Components.PersistedStateSerializationMode
 Microsoft.AspNetCore.Components.PersistedStateSerializationMode.Infer = 1 -> Microsoft.AspNetCore.Components.PersistedStateSerializationMode
 Microsoft.AspNetCore.Components.PersistedStateSerializationMode.Server = 2 -> Microsoft.AspNetCore.Components.PersistedStateSerializationMode
 Microsoft.AspNetCore.Components.PersistedStateSerializationMode.WebAssembly = 3 -> Microsoft.AspNetCore.Components.PersistedStateSerializationMode
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.FormMappingUseCurrentCulture.get -> bool
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.FormMappingUseCurrentCulture.set -> void
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.MaxFormMappingCollectionSize.get -> int
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.MaxFormMappingCollectionSize.set -> void
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.MaxFormMappingErrorCount.get -> int
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.MaxFormMappingErrorCount.set -> void
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.MaxFormMappingKeySize.get -> int
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.MaxFormMappingKeySize.set -> void
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.MaxFormMappingRecursionDepth.get -> int
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.MaxFormMappingRecursionDepth.set -> void
-Microsoft.Extensions.DependencyInjection.RazorComponentOptions.RazorComponentOptions() -> void
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.FormMappingUseCurrentCulture.get -> bool
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.FormMappingUseCurrentCulture.set -> void
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.MaxFormMappingCollectionSize.get -> int
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.MaxFormMappingCollectionSize.set -> void
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.MaxFormMappingErrorCount.get -> int
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.MaxFormMappingErrorCount.set -> void
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.MaxFormMappingKeySize.get -> int
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.MaxFormMappingKeySize.set -> void
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.MaxFormMappingRecursionDepth.get -> int
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.MaxFormMappingRecursionDepth.set -> void
+Microsoft.Extensions.DependencyInjection.RazorComponentsOptions.RazorComponentsOptions() -> void
 Microsoft.Extensions.DependencyInjection.RazorComponentsServiceCollectionExtensions
-static Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilderExtensions.AddComponentAssemblies(this Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder! builder, params System.Reflection.Assembly![]! assemblies) -> Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder!
-static Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilderExtensions.ConfigureComponentAssemblies(this Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder! builder, System.Action<Microsoft.AspNetCore.Components.Discovery.ComponentApplicationBuilder!>! configure) -> Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder!
-static Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilderExtensions.RemoveComponentAssemblies(this Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder! builder, params System.Reflection.Assembly![]! assemblies) -> Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder!
+static Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilderExtensions.AddAssemblies(this Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder! builder, params System.Reflection.Assembly![]! assemblies) -> Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder!
 static Microsoft.AspNetCore.Builder.RazorComponentsEndpointRouteBuilderExtensions.MapRazorComponents<TRootComponent>(this Microsoft.AspNetCore.Routing.IEndpointRouteBuilder! endpoints) -> Microsoft.AspNetCore.Builder.RazorComponentEndpointConventionBuilder!
-static Microsoft.Extensions.DependencyInjection.RazorComponentsServiceCollectionExtensions.AddRazorComponents(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services) -> Microsoft.AspNetCore.Components.Endpoints.IRazorComponentsBuilder!
-static Microsoft.Extensions.DependencyInjection.RazorComponentsServiceCollectionExtensions.AddRazorComponents(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, System.Action<Microsoft.Extensions.DependencyInjection.RazorComponentOptions!>! setupAction) -> Microsoft.AspNetCore.Components.Endpoints.IRazorComponentsBuilder!
+static Microsoft.Extensions.DependencyInjection.RazorComponentsServiceCollectionExtensions.AddRazorComponents(this Microsoft.Extensions.DependencyInjection.IServiceCollection! services, System.Action<Microsoft.Extensions.DependencyInjection.RazorComponentsOptions!>? configure = null) -> Microsoft.AspNetCore.Components.Endpoints.IRazorComponentsBuilder!
 static readonly Microsoft.AspNetCore.Components.Endpoints.RazorComponentResultExecutor.DefaultContentType -> string!
 virtual Microsoft.AspNetCore.Components.Endpoints.RazorComponentResultExecutor.ExecuteAsync(Microsoft.AspNetCore.Http.HttpContext! httpContext, Microsoft.AspNetCore.Components.Endpoints.RazorComponentResult! result) -> System.Threading.Tasks.Task!
