+++ src/Components/WebAssembly/Server/src/PublicAPI.Unshipped.txt
@@ -1,4 +1,4 @@
 #nullable enable
 Microsoft.AspNetCore.Components.WebAssembly.Server.TargetPickerUi.DisplayFirefox(Microsoft.AspNetCore.Http.HttpContext! context) -> System.Threading.Tasks.Task!
-Microsoft.Extensions.DependencyInjection.RazorComponentsBuilderExtensions
-static Microsoft.Extensions.DependencyInjection.RazorComponentsBuilderExtensions.AddWebAssemblyComponents(this Microsoft.AspNetCore.Components.Endpoints.IRazorComponentsBuilder! builder, System.Action<Microsoft.AspNetCore.Components.Endpoints.WebAssemblyComponentsEndpointOptions!>? configure = null) -> Microsoft.AspNetCore.Components.Endpoints.IRazorComponentsBuilder!
+Microsoft.Extensions.DependencyInjection.WebAssemblyRazorComponentsBuilderExtensions
+static Microsoft.Extensions.DependencyInjection.WebAssemblyRazorComponentsBuilderExtensions.AddWebAssemblyComponents(this Microsoft.AspNetCore.Components.Endpoints.IRazorComponentsBuilder! builder) -> Microsoft.AspNetCore.Components.Endpoints.IRazorComponentsBuilder!
